fix duplicate project name error for svn.xml and xml.xml

+++ rhino/io.js
@@ -1,17 +1,33 @@
 importPackage(java.io);
+
+load(__SCRIPTS_BASE__ + "/rhino/file_utils.js");
  
 var writeFile = function(string, file, encoding ) {
 	if(encoding == null || encoding == ""){
 		encoding = "UTF-8";
 	}
 	
-	var out = new BufferedWriter(
-		new OutputStreamWriter(
-			new FileOutputStream(file), encoding
-		)
-	);
+	var path = file;
+	file = new File(path);
+	if(!file.exists()){
+		var dir = FileUtils.dirname(path);
+		print(dir);
+		if(FileUtils.mkdirs(dir)){
+			file.createNewFile();
+		} else {
+			print("Error to create new file.");
+		}
+	}
 	
-	out.write(string);
-	out.flush();
-	out.close();
+	if(file.exists()){
+		var out = new BufferedWriter(
+			new OutputStreamWriter(
+				new FileOutputStream(file), encoding
+			)
+		);
+		
+		out.write(string);
+		out.flush();
+		out.close();
+	}
 }
